separated zcash metrics and blockchain explorer monitoring in different file and changed exception handling method
